Conflicts:
	ProjetoCruzeirao/.classpath
	ProjetoCruzeirao/pom.xml
	ProjetoCruzeirao/src/main/java/sistema/dados/Dados.java
	ProjetoCruzeirao/src/main/java/sistema/dao/UsuarioDAO.java
	ProjetoCruzeirao/src/main/java/sistema/daogeneric/DAOImpl.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Campeonato.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Cartao.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Categoria.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Equipe.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Fase.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Gol.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Grupo.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Inscricao.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Inscrito.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Juiz.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Local.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Partida.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/PartidasFutebol.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Rodada.java
	ProjetoCruzeirao/src/main/java/sistema/entidades/Usuario.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/CampeonatoMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/CartaoMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/CategoriaMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/EquipeMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/FaseMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/GolMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/GrupoMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/InscricaoMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/InscritoMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/JuizMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/LocalMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/PartidaFutebolMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/PartidaMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/RodadaMB.java
	ProjetoCruzeirao/src/main/java/sistema/managedbeans/UsuarioMB.java
	ProjetoCruzeirao/src/main/java/sistema/services/CampeonatoService.java
	ProjetoCruzeirao/src/main/java/sistema/services/CartaoService.java
	ProjetoCruzeirao/src/main/java/sistema/services/CategoriaService.java
	ProjetoCruzeirao/src/main/java/sistema/services/EquipeService.java
	ProjetoCruzeirao/src/main/java/sistema/services/FaseService.java
	ProjetoCruzeirao/src/main/java/sistema/services/GolService.java
	ProjetoCruzeirao/src/main/java/sistema/services/GrupoService.java
	ProjetoCruzeirao/src/main/java/sistema/services/InscricaoService.java
	ProjetoCruzeirao/src/main/java/sistema/services/InscritoService.java
	ProjetoCruzeirao/src/main/java/sistema/services/JuizService.java
	ProjetoCruzeirao/src/main/java/sistema/services/LocalService.java
	ProjetoCruzeirao/src/main/java/sistema/services/PartidaFutebolService.java
	ProjetoCruzeirao/src/main/java/sistema/services/PartidaService.java
	ProjetoCruzeirao/src/main/java/sistema/services/RodadaService.java
	ProjetoCruzeirao/src/main/java/sistema/services/UsuarioService.java
	ProjetoCruzeirao/src/main/webapp/acessonegado.xhtml
	ProjetoCruzeirao/src/main/webapp/login.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraCampeonato.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraCategoria.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraDadosPartida.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraEquipe.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraFase.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraGrupo.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraInscricao.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraInscrito.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraJuiz.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraLocal.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/cadastraRodada.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosCampeonatos.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosCategorias.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosEquipes.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosFases.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosGrupos.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosInscricoes.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosInscritos.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosJuizes.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosLocais.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosRodadas.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/dadosUsuarios.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/geracaoGrupos.xhtml
	ProjetoCruzeirao/src/main/webapp/pages/geracaoPartidas.xhtml

+++ ProjetoCruzeirao/src/main/java/sistema/managedbeans/CategoriaMB.java
@@ -1,36 +1,122 @@
 package sistema.managedbeans;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 
+import javax.faces.application.FacesMessage;
 import javax.faces.bean.ManagedBean;
-import javax.faces.bean.SessionScoped;
+import javax.faces.bean.ViewScoped;
+import javax.faces.context.FacesContext;
+
+import org.primefaces.event.RowEditEvent;
 
 import sistema.entidades.Categoria;
 import sistema.entidades.Fase;
 import sistema.entidades.Inscricao;
 import sistema.entidades.Inscrito;
+import sistema.services.CampeonatoService;
 import sistema.services.CategoriaService;
+import sistema.services.InscricaoService;
 
-@SessionScoped
+@ViewScoped
 @ManagedBean
 public class CategoriaMB {
+	
 	private Categoria categoria = new Categoria();
 	private CategoriaService categoriaservice = new CategoriaService();
 	private Inscricao auxInscricao;
 	private Fase auxFase;
+	private List<Categoria> categorias;
+	private Categoria auxCat;
+	private List<Inscricao> inscricoes = new ArrayList<Inscricao>();
+	private InscricaoService inscricaoservice = new InscricaoService();
+	private CampeonatoService campeonatoservice = new CampeonatoService();
 	
-	
+	public void onRowEdit(RowEditEvent event) {
+
+		Categoria c = ((Categoria) event.getObject());
+		categoriaservice.alterar(c);
+	}
 
 	public void salvar() {
-		categoria.setIdCategoria(0);
-		categoriaservice.salvar(categoria);
-		categoria.getCampeonato().getCategorias().add(categoria);
 		
-		for (Inscricao aux : categoria.getInscricoes())		
-			for(Inscrito aux2 : aux.getInscritos())
-				aux2.getUsuario().getCampeonatos().add(categoria.getCampeonato());
+		int count = 0;		
+		
+		
+		if (this.getCategorias() != null)
+			for (Categoria c : this.getCategorias()) 
+				if (c.getNome().equals(categoria.getNome()))
+					count++;
+			
+		if (categoria.getMinJogadores() >= categoria.getMaxJogadores()) 
+		{
+			FacesMessage mensagem = new FacesMessage("O número mínimo de jogadores não pode ser maior ou igual ao número máximo!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+		}
+		
+		else if (count > 0) 
+		{
+			FacesMessage mensagem = new FacesMessage("Já existe uma categoria com esse nome!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+		}
+		
+		else
+		{
+
+			if (categoria.getCampeonato() != null) {
+				categoria.getCampeonato().getCategorias().add(categoria);
+				//campeonatoservice.alterar(categoria.getCampeonato());
+			}
+				
+			
+			categoria = categoriaservice.salvar(categoria);
+			
+			if (categorias != null) {
+				categorias.add(categoria);
+					
+				
+				FacesMessage mensagem = new FacesMessage("Categoria " + categoria.getNome() + " cadastrada com sucesso!");
+				FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+				
+				/*for (Inscricao aux : categoria.getInscricoes())		
+					for (Inscrito aux2 : aux.getInscritos())
+						aux2.getUsuario().getCampeonatos().add(categoria.getCampeonato());*/
+				
+			}
+			
+		}
 
 		categoria = new Categoria();
+	}
+	
+	public void atualizarInscricoes() {
+		for (Inscricao i : auxCat.getInscricoes()) {
+			if (!i.isValidada())
+				inscricoes.add(i);
+		}
+	}
+	
+	public void validarInscricao() {
+		int count = 0;
+		if ((auxInscricao.getInscritos().size() >= auxInscricao.getCategoria().getMinJogadores()) && (auxInscricao.getInscritos().size() <= auxInscricao.getCategoria().getMaxJogadores())) {			// Quantidade minima e maxima de jogadores
+			for (Inscrito inscritosatual : auxInscricao.getInscritos()) {
+				/*if (inscritosatual.getUsuario().getDataNascimento().getYear() < auxCat.getNascidosAPartirDe())
+					count++;
+				*/
+				for (Inscricao i : auxCat.getInscricoes()) {
+					for (Inscrito inscritos : i.getInscritos()) {
+						if (inscritosatual.getUsuario() == inscritos.getUsuario())
+							count++;
+					}
+				}
+			}
+			
+			if (count == 0) {
+				auxInscricao.setValidada(true);
+				inscricoes.remove(auxInscricao);
+				inscricaoservice.alterar(auxInscricao);
+			}
+		}
 	}
 	
 	public void adicionarInscricao() {
@@ -65,9 +151,24 @@ public class CategoriaMB {
 		this.auxInscricao = auxInscricao;
 	}
 
-	public void remover(Categoria categoria) {
-		categoriaservice.remover(categoria);
-		categoria.getCampeonato().getCategorias().remove(categoria);
+	public void remover(Categoria categoria) {		
+		
+		if (categoria.getCampeonato() != null || categoriaservice.pesquisarFasesCategoria(categoria).size() > 0 || categoriaservice.pesquisarInscricoesCategoria(categoria).size() > 0) {
+			FacesContext context = FacesContext.getCurrentInstance();
+			context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Não é possível remover a categoria, pois há equipe (es), ou campeonato (os) ou inscrição (ões) amarradas a ele.", null));
+		}
+		
+		else {
+			categoriaservice.remover(categoria);
+			categorias.remove(categoria);
+			
+			if (categoria.getCampeonato() != null)
+				categoria.getCampeonato().getCategorias().remove(categoria);
+			
+			FacesMessage mensagem = new FacesMessage("Categoria " + categoria.getNome() + " removida com sucesso!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);
+		}
+		
 	}
 
 	public Categoria getCategoria() {
@@ -78,7 +179,26 @@ public class CategoriaMB {
 		this.categoria = categoria;
 	}
 	
-	public ArrayList<Categoria> getCategorias() {
-		return categoriaservice.getCategorias();
+	public List<Categoria> getCategorias() {
+		if (categorias == null)
+			categorias = categoriaservice.getCategorias();
+		
+		return categorias;
+	}
+
+	public Categoria getAuxCat() {
+		return auxCat;
+	}
+
+	public void setAuxCat(Categoria auxCat) {
+		this.auxCat = auxCat;
+	}
+
+	public List<Inscricao> getInscricoes() {
+		return inscricoes;
+	}
+
+	public void setInscricoes(List<Inscricao> inscricoes) {
+		this.inscricoes = inscricoes;
 	}
 }
