+++ ProjetoCruzeirao/src/main/java/sistema/managedbeans/InscritoMB.java
@@ -1,29 +1,65 @@
 package sistema.managedbeans;
 
-import java.util.*;
+import java.io.Serializable;
+import java.util.List;
 
+import javax.faces.application.FacesMessage;
 import javax.faces.bean.ManagedBean;
-import javax.faces.bean.SessionScoped;
+import javax.faces.bean.ViewScoped;
+import javax.faces.context.FacesContext;
+
+import org.primefaces.event.RowEditEvent;
 
 import sistema.entidades.Inscrito;
 import sistema.services.InscritoService;
 
-@SessionScoped
+@ViewScoped
 @ManagedBean
-public class InscritoMB {
+public class InscritoMB implements Serializable{
+	
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
 	private Inscrito inscrito = new Inscrito();
 	private InscritoService inscritoservice = new InscritoService();
+	private List<Inscrito> inscritos;
+	
+	public void onRowEdit(RowEditEvent event) {
+
+		Inscrito i = ((Inscrito) event.getObject());
+		inscritoservice.alterar(i);
+	}
 	
 	public void salvar() {
-		inscrito.setIdInscrito(0);
-		inscritoservice.salvar(inscrito);
+		
 		inscrito.getInscricao().getInscritos().add(inscrito);
-		inscrito.getUsuario().getInscricoes().add(inscrito);
+		inscrito.getUsuario().getInscricoes().add(inscrito);		
+		inscrito = inscritoservice.salvar(inscrito);
+		
+		if (inscritos != null) {
+			inscritos.add(inscrito);	
+			FacesMessage mensagem = new FacesMessage("Inscrito " + inscrito.getUsuario().getNome() + " cadastrado com sucesso!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);			
+		}
+		
 		inscrito = new Inscrito();
 	}
 	
 	public void remover(Inscrito inscrito) {
-		inscritoservice.remover(inscrito);
+		
+		if (inscrito.getInscricao() != null || inscrito.getUsuario() != null) {
+			FacesContext context = FacesContext.getCurrentInstance();
+			context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Não é possível remover o inscrito, pois há inscrição ou usuário amarrado a ele.", null));
+		}
+		
+		else {
+			inscritoservice.remover(inscrito);
+			inscritos.remove(inscrito);
+			FacesMessage mensagem = new FacesMessage("Inscrito " + inscrito.getUsuario().getNome() + " removido com sucesso!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);
+		}
+				
 	}
 
 	public Inscrito getInscrito() {
@@ -34,8 +70,11 @@ public class InscritoMB {
 		this.inscrito = inscrito;
 	}
 	
-	public ArrayList<Inscrito> getInscritos() {
-		return inscritoservice.getInscritos();
+	public List<Inscrito> getInscritos() {
+		if (inscritos == null)
+			inscritos = inscritoservice.getInscritos();
+	
+		return inscritos;
 	}
 	
 }
